Code cleanup for rendering of dynamic hitbox shader
